refactor_follow_function

+++ db/migrate/20251028073217_create_follows.rb
@@ -1,8 +1,8 @@
 class CreateFollows < ActiveRecord::Migration[8.1]
   def change
     create_table :follows do |t|
-      t.integer :follower_id
-      t.integer :followed_id
+      t.bigint :follower_id
+      t.bigint :followed_id
 
       t.timestamps
     end
